+++ src/IDInterface.java
@@ -2,5 +2,5 @@ interface IDInterface {
 	/**
 	 * This function generates a random ID.
 	 */
-	int get(int length);
+	long get(int length);
 }
